move common CRDT flow types to types.js
